refactor: extract DiagnosticsLogCategory.swift from DiagnosticsTypes.swift
Splits the `DiagnosticsLogCategory` enum (the CaseIterable domain
categories) out of DiagnosticsTypes.swift.

Byte-identical move of lines 4-12. Only imports Foundation (no OSLog
needed for the plain string enum).

DiagnosticsTypes.swift: 92 → 82 lines. Tests: 667.
